feat: refactor ollama app to separate out ollama api calls into a separate file
